Maven化. plugin.ymlに関してはとりあえず仮置き.
