+++ src/main/resources/plugin.yml
@@ -1,0 +1,4 @@
+main: jp.kotmw.loginbonus.Main
+name: LoginBonusPlugin
+author: kotmw0701
+version: 0.0.1
